+++ tests/Integration/Services/Task/Service/TaskChatTest.php
@@ -13,46 +13,54 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Tests\Integration\Services\Task\Service;
 
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Exceptions\TransportException;
+use Bitrix24\SDK\Core;
 use Bitrix24\SDK\Infrastructure\Filesystem\Base64Encoder;
 use Bitrix24\SDK\Services\ServiceBuilder;
+use Bitrix24\SDK\Services\Task\Result\TaskItemResult;
 use Bitrix24\SDK\Services\Task\Service\Task;
+use Bitrix24\SDK\Services\Task\Service\TaskChat;
 use Bitrix24\SDK\Services\Task\Service\TaskFile;
 use Bitrix24\SDK\Services\Task\Service\TaskItemBuilder;
+use Bitrix24\SDK\Services\Task\Service\TaskItemSelectBuilder;
+use Bitrix24\SDK\Services\User\Service\User;
+use Bitrix24\SDK\Tests\Builders\Services\Task\TaskBuilder;
 use Bitrix24\SDK\Tests\CustomAssertions\CustomBitrix24Assertions;
 use Bitrix24\SDK\Tests\Integration\Factory;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\CoversFunction;
 use PHPUnit\Framework\Attributes\CoversMethod;
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
 use Symfony\Component\Filesystem\Filesystem;
 
-#[CoversMethod(TaskFile::class, 'attachExists')]
-class TaskFileTest extends TestCase
+/**
+ * Class TaskTest
+ *
+ * @package Bitrix24\SDK\Tests\Integration\Services\Task\Service
+ */
+#[CoversMethod(TaskChat::class, 'sendMessage')]
+class TaskChatTest extends TestCase
 {
     use CustomBitrix24Assertions;
 
     protected Task $taskService;
-    protected TaskFile $taskFileService;
+    protected TaskChat $taskChatService;
 
     protected ServiceBuilder $serviceBuilder;
-
-    protected Base64Encoder $base64Encoder;
 
     #[\Override]
     protected function setUp(): void
     {
         $this->taskService = Factory::getServiceBuilder(false)->getTaskScope()->task();
-        $this->taskFileService = Factory::getServiceBuilder(false)->getTaskScope()->taskFile();
+        $this->taskChatService = Factory::getServiceBuilder(false)->getTaskScope()->taskChat();
         $this->serviceBuilder = Factory::getServiceBuilder();
-        $this->base64Encoder = new Base64Encoder(
-            new Filesystem(),
-            new \Symfony\Component\Mime\Encoder\Base64Encoder(),
-            new NullLogger()
-        );
     }
 
-    #[TestDox('Upload existing file to task')]
-    public function uploadExistingFileToTask(): void
+    #[TestDox('Send message to task chat')]
+    public function testGetTaskByIdWithAllFields(): void
     {
         $curUser = $this->serviceBuilder->getUserScope()->user()->current()->user();
         $addedTask = $this->taskService->add(
@@ -63,26 +71,7 @@ class TaskFileTest extends TestCase
             )
         );
 
-        $rootStorageId = (int)$this->serviceBuilder->getDiskScope()->storage()->list(
-            [
-                'ID' => $curUser->ID,
-                'ENTITY_TYPE' => 'user'
-            ]
-        )->storages()[0]->ID;
-
-        $testContent = 'Test file content - ' . time();
-        $base64Content = $this->base64Encoder->encodeString($testContent);
-        $fileData = [
-            'NAME' => 'test_file_' . time() . '.txt'
-        ];
-        $uploadedFileResult = $this->serviceBuilder->getDiskScope()->folder()->uploadFile(
-            $rootStorageId,
-            $fileData,
-            $base64Content,
-            true
-        );
-
-        $this->assertTrue($this->taskFileService->attachExists($addedTask->task()->id, [$uploadedFileResult->getId()])->isSuccess());
+        $this->assertTrue($this->taskChatService->sendMessage($addedTask->task()->id, 'Hello world')->isSuccess());
         $this->taskService->delete($addedTask->task()->id);
     }
 }